+++ packages/rum-core/src/rawRumEvent.types.ts
@@ -242,6 +242,9 @@ export interface RumContext {
     session: {
       plan: RumSessionPlan
     }
+    view?: {
+      document_version: number
+    }
     browser_sdk_version?: string
   }
 }